refactoring_backups/snapshots/baseline_20250618_140631/orchestrator/requirements-dev.txt
# Fichier : orchestrator/requirements-dev.txt
# Dépendances de développement et test

# --- Tests ---
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-html==4.1.1

# --- Sécurité & Static Analysis ---
bandit==1.7.5
safety==2.3.5
semgrep==1.45.0
mypy==1.8.0

# --- Formatage & Linting ---
black==23.12.1
isort==5.13.2
flake8==6.1.0
pre-commit==3.6.0

# --- Observabilité avancée ---
structlog==23.2.0
python-json-logger==2.0.7

# --- Tests de charge ---
locust==2.20.0

# --- Mock & Fixtures ---
factory-boy==3.3.0
responses==0.24.1
httpx-mock==0.10.1
